[Section 2] Organize view controllers
